requirements.txt
flask==2.0.2
pandas~=1.3.4
numpy~=1.21.4
matplotlib~=3.5.1
Pillow~=9.0.0
flask-restful
flask-cors